Renames `ValueGenerator` to `Distribution`
Simplifies corresponding case class names.

All distributions produce `Double` which could make things difficult
if you need discrete values. Need to find a solution to that.

Task cost is now a double as a result of that.

+++ proter/src/main/scala/com/workflowfm/proter/TaskResource.scala
@@ -26,7 +26,7 @@ import java.util.UUID
   * @param name The name of the resource.
   * @param costPerTick The cost of using the resource per unit of time.
   */
-class TaskResource(val name: String, val costPerTick: Int) {
+class TaskResource(val name: String, val costPerTick: Double) {
   /**
     * Some [[TaskInstance]] currently attached to the resource, or [[scala.None]] if idle.
     */
